MyApps-13: Move cypher implementation, api and dependency injection to submodules.
