+++ data/di/src/main/java/xyz/torquato/myapps/data/di/RepositoryImpl.kt
@@ -4,13 +4,19 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import xyz.torquato.myapps.data.impl.cypher.CypherRepository
 import xyz.torquato.myapps.data.impl.paging.PagingRepository
+import xyz.torquato.myaps.domain.api.encrypt.ICypherRepository
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryImpl {
+
+    @Binds
+    @Singleton
+    abstract fun bindCypherImpl(impl: CypherRepository): ICypherRepository
 
     @Binds
     @Singleton
