+++ data/impl/src/androidTest/java/xyz/torquato/myapps/data/impl/CypherTester.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps
+package xyz.torquato.myapps.data.impl
 
 object CypherTester {
