+++ data/impl/src/main/cpp/CMakeLists.txt
@@ -5,8 +5,4 @@
 # Sets the minimum CMake version required for this project.
 cmake_minimum_required(VERSION 3.22.1)
 
-add_subdirectory(basic)
-add_subdirectory(waves)
-add_subdirectory(myengine)
-add_subdirectory(web)
-add_subdirectory(calculator)
+add_subdirectory(cypher)
