+++ data/impl/src/main/cpp/cypher/inc/slot.h
