+++ data/impl/src/main/cpp/cypher/src/impl/infmath.cpp
