+++ data/impl/src/main/cpp/cypher/src/impl/slot.cpp
