+++ data/impl/src/main/cpp/cypher/src/utils/CypherUtil.hpp
