+++ data/impl/src/main/cpp/cypher/src/utils/StringUtil.hpp
