+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/cypher/CypherRepository.kt
@@ -1,11 +1,9 @@
-package xyz.torquato.myapps.data.cypher
+package xyz.torquato.myapps.data.impl.cypher
 
-import androidx.annotation.VisibleForTesting
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import xyz.torquato.myapps.api.sound.ICypherRepository
+import xyz.torquato.myaps.domain.api.encrypt.ICypherRepository
 import javax.inject.Inject
-import kotlin.toUByte
 
 class CypherRepository @Inject constructor() : ICypherRepository {
 
@@ -17,8 +15,6 @@ class CypherRepository @Inject constructor() : ICypherRepository {
     override val message = _message.asStateFlow()
 
     private external fun entry(token: String, message: String): String
-
-
 
 
     override fun setMessage(newMessage: String) {
